requirements.txt
requests 
python-dotenv 
pymysql 
